folding::expressions

Enum FoldingCompatibleExprInner

pub enum FoldingCompatibleExprInner<C: FoldingConfig> {
    Constant(<C::Curve as AffineRepr>::ScalarField),
    Challenge(C::Challenge),
    Cell(Variable<C::Column>),
    Extensions(ExpExtension<C>),
}

Components to be used to convert multivariate polynomials into โcompatibleโ multivariate polynomials that will be translated to folding expressions.

Variants

Constant(<C::Curve as AffineRepr>::ScalarField)

Challenge(C::Challenge)

Cell(Variable<C::Column>)

Extensions(ExpExtension<C>)

extra nodes created by folding, should not be passed to folding