folding

Trait FoldingEnv

pub trait FoldingEnv<F: Zero + Clone, I, W, Col, Chal, Selector> {
    type Structure;

    // Required methods
    fn new(
        structure: &Self::Structure,
        instances: [&I; 2],
        witnesses: [&W; 2],
    ) -> Self;
    fn challenge(&self, challenge: Chal, side: Side) -> F;
    fn col(&self, col: Col, curr_or_next: CurrOrNext, side: Side) -> &[F];
    fn selector(&self, s: &Selector, side: Side) -> &[F];
}

Describe a folding environment. The type parameters are:

F: The field of the circuit/computation

I: The instance type, i.e the public inputs

W: The type of the witness, i.e. the private inputs

Col: The type of the column

Chal: The type of the challenge

Selector: The type of the selector

Required Associated Types

type Structure

Structure which could be storing useful information like selectors, etc.

Required Methods

fn new( structure: &Self::Structure, instances: [&I; 2], witnesses: [&W; 2], ) -> Self

Creates a new environment storing the structure, instances and witnesses.

fn challenge(&self, challenge: Chal, side: Side) -> F

Obtains a given challenge from the expanded instance for one side. The challenges are stored inside the instances structs.

fn col(&self, col: Col, curr_or_next: CurrOrNext, side: Side) -> &[F]

Returns the evaluations of a given column witness at omega or zeta*omega.

fn selector(&self, s: &Selector, side: Side) -> &[F]

similar to Self::col, but folding may ask for a dynamic selector directly instead of just column that happens to be a selector

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.