Enum folding::expressions::FoldingCompatibleExpr

pub enum FoldingCompatibleExpr<C: FoldingConfig> {
    Atom(FoldingCompatibleExprInner<C>),
    Pow(Box<Self>, u64),
    Add(Box<Self>, Box<Self>),
    Sub(Box<Self>, Box<Self>),
    Mul(Box<Self>, Box<Self>),
    Double(Box<Self>),
    Square(Box<Self>),
}

Compatible folding expressions that can be used with folding schemes. An expression from kimchi::circuits::expr::Expr can be converted into a FoldingCompatibleExpr using the trait From. From there, an expression of type IntegratedFoldingExpr can be created using the function folding_expression.

Variants

Atom(FoldingCompatibleExprInner<C>)

Pow(Box<Self>, u64)

Add(Box<Self>, Box<Self>)

Sub(Box<Self>, Box<Self>)

Mul(Box<Self>, Box<Self>)

Double(Box<Self>)

Square(Box<Self>)

Implementations

impl<C: FoldingConfig> FoldingCompatibleExpr<C>

Converts an expression “compatible” with folding into a folded expression.

pub fn simplify(self) -> FoldingExp<C>

pub fn map_variable( self, mapper: &dyn Fn(Variable<C::Column>) -> Variable<C::Column> ) -> FoldingCompatibleExpr<C>

Maps variable (column index) in expression using the mapper function. Can be used to modify (remap) the indexing of columns after the expression is built.

pub fn flatten_quad_columns( self, mapper: &dyn Fn(usize) -> Variable<C::Column> ) -> FoldingCompatibleExpr<C>

Map all quad columns into regular witness columns.